feat(node): Use `process.on('SIGTERM')` for flushing in Vercel functions (#17583)
This adds a better way to ensure we flush in node-based vercel
functions.

Note that this _does not_ work in edge functions, so we need to keep the
`vercelWaitUntil` code around for this - but we can noop it in node and
instead rely on the better way to handle this, I believe.

@chargome when you're back, can you verify if this makes sense? 😅 

Closes https://github.com/getsentry/sentry-javascript/issues/17567

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>