ref(core): Extract and reuse `getCombinedScopeData` helper (#18585)
Pre-work for #18160

We have multiple places in which we need to combine the `ScopeData` of
global, isolation and current scopes, to apply this data to telemetry
items (events, logs, metrics, soon also spansV2). Previously, we did
this in-place or with helpers that were not re-used. This PR now unifies
the various locations to one helper from core which can be reused
everywhere.

Closes #18586 (added automatically)