fix(feedback): Fix cases where the outline of inputs were wrong
When the integration had "colorScheme: "dark"," then we include css for
both light and dark mode. But some properties, like `outline` don't need
to be specified, the system does it for us. However, setting
`color-scheme: only light` meant that the light-mode outlines were used
in all cases.

This changes things so that we have the correct values for
`color-scheme` if a specific value is picked. This ensures that the
`outline` and `:focus` colors set by the system are correct in all cases.

Test Plan:
I tested with my system set to each of: light, dark, automatic
And then with the integration setting set to each of: `colorScheme:
'light'`, `colorScheme: 'dark'` and `colorScheme: 'system'`.
To test i just opened up the feedback widget and clicked to focus an
input box.